added token validation

+++ ugc_2/src/api/v1/reviews.py
@@ -2,6 +2,7 @@ from datetime import datetime
 from http import HTTPStatus
 
 from fastapi import APIRouter, Depends, Request, HTTPException
+from fastapi_jwt_auth import AuthJWT
 from models.events import Review, ReviewPosted
 from services.reviews import (ReviewsService, get_events_service,
                               review_serializer)
@@ -18,12 +19,17 @@ router = APIRouter()
 async def add_review(
     request: Request,
     event: Review,
+    Authorize: AuthJWT = Depends(),
     service: ReviewsService = Depends(get_events_service),
 ):
     logger.info(f"request add review {request}")
+    Authorize.fresh_jwt_required()
+    user_id = Authorize.get_jwt_subject()
     try:
         new_review = await service.add_event(
-            ReviewPosted(**event.dict(), created_at=datetime.now())
+            ReviewPosted(**event.dict(),
+                        created_at=datetime.now(),
+                        user_id=user_id)
         )
     except DocumentExistsException:
         raise HTTPException(status_code=HTTPStatus.CONFLICT)
@@ -36,9 +42,17 @@ async def update_review(
     request: Request,
     review_id: str,
     event: Review,
+    Authorize: AuthJWT = Depends(),
     service: ReviewsService = Depends(get_events_service),
 ):
     logger.info(f"request update review {request}")
+    Authorize.fresh_jwt_required()
+    user_id = Authorize.get_jwt_subject()
+    review = await service.find_one(review_id)
+    if not review:
+        return HTTPStatus.NOT_FOUND
+    if review.get('user_id') != user_id:
+        return HTTPStatus.FORBIDDEN
     updated_review = await service.update(review_id, event)
     review = review_serializer(updated_review)
     return review
@@ -48,8 +62,16 @@ async def update_review(
 async def delete_review(
     request: Request,
     review_id: str,
+    Authorize: AuthJWT = Depends(),
     service: ReviewsService = Depends(get_events_service),
 ):
     logger.info(f"request delete review {request}")
-    result = await service.delete(review_id)
-    return HTTPStatus.NO_CONTENT if result else HTTPStatus.NOT_FOUND
+    Authorize.fresh_jwt_required()
+    user_id = Authorize.get_jwt_subject()
+    review = await service.find_one(review_id)
+    if not review:
+        return HTTPStatus.NOT_FOUND
+    if review.get('user_id') != user_id:
+        return HTTPStatus.FORBIDDEN
+    await service.delete(review_id)
+    return HTTPStatus.NO_CONTENT
